requirements.txt
# Core OCR libraries
pytesseract==0.3.13
Pillow==11.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
easyocr==1.7.2

# Image processing and enhancement
numpy==2.2.6
scikit-image==0.25.2
imutils==0.5.4

# PDF processing
PyPDF2==3.0.1
pdf2image==1.17.0

# Machine Learning (for advanced OCR)
torch==2.8.0
torchvision==0.23.0

# Web interface
flask==3.1.2
streamlit==1.50.0

# Utilities
python-dotenv==1.1.1
click==8.3.0
tqdm==4.67.1

# Dependencies
scipy==1.16.2
pandas==2.3.3
requests==2.32.5
pyyaml==6.0.3
shapely==2.1.2
pyclipper==1.3.0.post6
python-bidi==0.6.6
tifffile==2025.10.4
imageio==2.37.0
lazy_loader==0.4
ninja==1.13.0
filelock==3.19.1
typing_extensions==4.15.0
setuptools==80.9.0
sympy==1.14.0
mpmath==1.3.0
networkx==3.5
jinja2==3.1.6
markupsafe==3.0.3
fsspec==2025.9.0
packaging==25.0
altair==5.5.0
cachetools==6.2.0
protobuf==6.32.1
pyarrow==21.0.0
pydeck==0.9.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
urllib3==2.5.0
werkzeug==3.1.3
blinker==1.9.0
itsdangerous==2.2.0
gitpython==3.1.45
gitdb==4.0.12
smmap==5.0.2
attrs==25.4.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
charset-normalizer==3.4.3
idna==3.10
certifi==2025.10.5
narwhals==2.7.0
six==1.17.0

# Development tools (optional)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0